Replace the old parser with the new parser
