+++ lib/graphql/language.rb
@@ -8,7 +8,6 @@ require "graphql/language/lexer"
 require "graphql/language/nodes"
 require "graphql/language/cache"
 require "graphql/language/parser"
-require "graphql/language/recursive_descent_parser"
 require "graphql/language/static_visitor"
 require "graphql/language/token"
 require "graphql/language/visitor"
