+++ spec/graphql/language/parser_spec.rb
@@ -8,11 +8,7 @@ describe GraphQL::Language::Parser do
     err = assert_raises GraphQL::ParseError do
       subject.parse("{ foo(query: \"\xBF\") }")
     end
-    expected_message = if using_recursive_descent_parser?
-      'Parse error on bad Unicode escape sequence'
-    else
-      'Parse error on bad Unicode escape sequence: "{ foo(query: \"\xBF\") }" (error) at [1, 1]'
-    end
+    expected_message = 'Parse error on bad Unicode escape sequence'
     assert_equal expected_message, err.message
   end
 
@@ -34,10 +30,8 @@ describe GraphQL::Language::Parser do
     }
     expected_msg = if USING_C_PARSER
       "syntax error, unexpected invalid token (\"\\xF0\"), expecting LCURLY at [1, 7]"
-    elsif using_recursive_descent_parser?
-      "Expected LCURLY, actual: UNKNOWN_CHAR (\"\\xF0\") at [1, 7]"
     else
-      "Parse error on \"😘\" (error) at [1, 7]"
+      "Expected LCURLY, actual: UNKNOWN_CHAR (\"\\xF0\") at [1, 7]"
     end
 
     assert_equal expected_msg, err.message
@@ -361,15 +355,15 @@ GRAPHQL
     query = GraphQL::Query.new(schema, "{ t: __typename }")
     subject.parse("{ t: __typename }", trace: query.current_trace)
     traces = TestTracing.traces
-    expected_traces = if using_recursive_descent_parser?
-      1
-    else
+    expected_traces = if USING_C_PARSER
       2
+    else
+      1
     end
     assert_equal expected_traces, traces.length
     lex_trace, parse_trace = traces
 
-    if !using_recursive_descent_parser?
+    if USING_C_PARSER
       assert_equal "{ t: __typename }", lex_trace[:query_string]
       assert_equal "lex", lex_trace[:key]
       assert_instance_of Array, lex_trace[:result]
